+++ lib/src/core/manager.dart
@@ -6,6 +6,8 @@ part of dartemis;
 abstract class Manager implements EntityObserver {
   World _world;
 
+  World get world => _world;
+
   /**
    * Override to implement code that gets executed when managers are initialized.
    */